+++ src/shape/quad.rs
@@ -41,7 +41,7 @@ use serde::{de::DeserializeOwned, Deserialize, Serialize};
 #[must_use]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[cfg_attr(feature = "serde", serde(bound = "T: Serialize + DeserializeOwned"))]
-pub struct Quad<T, const N: usize>(pub(crate) [Point<T, N>; 4]);
+pub struct Quad<T = i32, const N: usize = 2>(pub(crate) [Point<T, N>; 4]);
 
 /// A 2D `Quad` represented by `i32`.
 pub type QuadI2 = Quad<i32, 2>;