+++ src/vector.rs
@@ -95,7 +95,7 @@ use std::{fmt, ops::MulAssign};
 #[must_use]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[cfg_attr(feature = "serde", serde(bound = "T: Serialize + DeserializeOwned"))]
-pub struct Vector<T, const N: usize>(
+pub struct Vector<T = Scalar, const N: usize = 2>(
     #[cfg_attr(feature = "serde", serde(with = "arrays"))] pub(crate) [T; N],
 );